Change generators/tests to use SourceOfRandomness enhanced with range methods

+++ src/main/java/com/pholser/junit/quickcheck/generator/ByteGenerator.java
@@ -46,6 +46,6 @@ public class ByteGenerator extends Generator<Byte> {
 
     @Override
     public Byte generate(SourceOfRandomness random, GenerationStatus status) {
-        return (byte) random.nextInt(min, max);
+        return random.nextByte(min, max);
     }
 }
